Archivos eliminados y componentes creados
ChatInfo: añade o quita el ChatManager.
ChestPanel: añade o quita ObjectChest.
