--- Proyecto.02/project/src/Components/Paneles/Cartel_Info.cpp
@@ -1,40 +1,0 @@
-#include "Cartel_Info.h"
-
-void CartelInfo::init()
-{
-}
-
-void CartelInfo::update()
-{
-}
-
-void CartelInfo::draw()
-{
-}
-
-void CartelInfo::drawCartelActual()
-{
-	// Fondo:
-	pan_->addButton(iManager_->addButton<SDL_Object>(marco, src::Cartel));
-
-	// TITLE:
-	addTitleName();
-}
-
-void CartelInfo::addTitleName()
-{
-	switch (type)
-	{
-	default:
-		break;
-	}
-}
-
-void CartelInfo::addTemplate()
-{
-}
-
-std::string CartelInfo::checkLineSize(std::string line)
-{
-    return std::string();
-}
